API reference

Basic schema

Functions to create basic Markov individual-based models (IBM).

individual.schema_base.TheoryIBM — Constant

ACSet definition for a basic individual-based model See Catlab.jl documentation for description of the @present syntax.

individual.schema_base.AbstractIBM — Type

An abstract ACSet for a basic Markov individual-based model inheriting from AbstractIBM which allows for events to be scheduled for persons.

Missing docstring.

Missing docstring for individual.schema_base.IBM. Check Documenter's build log for details.

individual.schema_base.npeople — Function

npeople(model::AbstractIBM, states)

Return the number of people in some set of `states` (an element of the State Ob).
If called without the argument `states`, simply return the total population size.

individual.schema_base.nstate — Function

nstate(model::AbstractIBM)

Return the size of the finite state space.

individual.schema_base.statelabel — Function

statelabel(model::AbstractIBM)

Return the labels (names) of the states in the finite state space.

individual.schema_base.get_index_state — Function

getindexstate(model::AbstractIBM, states)

Return an integer vector giving the persons who are in the states specified in `states`.
If called without the argument `states`, simply return everyone's index.

individual.schema_base.queue_state_update — Function

queuestateupdate(model::AbstractIBM, persons, state)

For persons specified in `persons`, queue a state update to `state`, which will be applied at the
end of the time step.

individual.schema_base.apply_state_updates — Function

applystateupdates(model::AbstractIBM)

Apply all queued state updates.

individual.schema_base.output_states — Function

output_states(t::Int, model::AbstractIBM)

Return a vector counting the number of persons in each state.

individual.schema_base.initialize_states — Function

initializestates(model::AbstractIBM, initialstates, state_labels::Vector{String})

Initialize the categorical states of a model. The argument `initial_states` can either
be provided as a vector of integers, corresponding to the internal storage of the ACSet,
or as a vector of strings. It should be equal in length to the population which is to be
simulated.

individual.schema_base.reset_states — Function

initializestates(model::AbstractIBM, initialstates)

Reset a model's categorical states.

Event scheduling schema

Create IBMs with event scheduling capabilities.

individual.schema_events.TheorySchedulingIBM — Constant

A schema for an individual-based model inheriting from TheoryIBM which allows for events to be scheduled for persons.

individual.schema_events.AbstractSchedulingIBM — Type

An abstract ACSet for an individual-based model inheriting from AbstractIBM which allows for events to be scheduled for persons.

Missing docstring.

Missing docstring for individual.schema_events.SchedulingIBM. Check Documenter's build log for details.

individual.schema_events.schedule_event — Function

schedule_event(model::AbstractSchedulingIBM, target, delay, event)

Schedule a set of persons in `target` for the `event` after some `delay`. Note that `event` should correspond to an element
in the set `EventLabel` in your model.

individual.schema_events.get_scheduled — Function

get_scheduled(model::AbstractSchedulingIBM, event)

Get the set of persons scheduled for `event`. Note that `event` should correspond to an element
    in the set `EventLabel` in your model.

individual.schema_events.clear_schedule — Function

clear_schedule(model::AbstractSchedulingIBM, target)

Clear the persons in `target` from any events they are scheduled for.
If 'target' is not specified, clear all scheduled events.

individual.schema_events.event_tick — Function

event_tick(model::AbstractSchedulingIBM)

Reduce all delays by 1, called at the end of a time step.

individual.schema_events.event_process — Function

event_process(model::AbstractSchedulingIBM, t::Int)

Process events which are ready to fire.

Sampling

Methods for sampling random variates commonly used in IBMs.

individual.sampling.delay_sample — Function

delay_sample(n::Integer, rate::AbstractFloat, dt::AbstractFloat)

Sample time steps until an event fires given a rate and dt.

individual.sampling.bernoulli_sample — Function

bernoulli_sample(target::AbstractVector, prob::AbstractFloat)

Sample without replacement from target with success probability prob.

bernoulli_sample(target::AbstractVector, rate::AbstractFloat, dt::AbstractFloat)

Sample without replacement from target with success probability calculated from $1 - e^{-\mathrm{rate} * \mathrm{dt}}$

bernoulli_sample(target::AbstractVector{T}, prob::Vector)

Sample without replacement from target where each element has a unique success probability given in the vector prob.

bernoulli_sample(target::AbstractVector{T}, rate::Vector, dt::AbstractFloat)

Sample without replacement from target where each element's success probability is calculated from $1 - e^{-\mathrm{rate} * \mathrm{dt}}$.

individual.sampling.choose — Function

choose(target::T, K::Integer)

Return a vector of size K with that number of random elements selected without replacement from target.